Fixed Opening text notification
the modded notifications were opening on top of the default contacts UI. This adds an unnoticeable delay to make sure that the modded UI opens properly. Also added recognition for the Escape key to fix a bug that left the modded menu open when you hit escape while having it open.